requirements.txt
asgiref==3.8.1
dj-database-url==3.0.1
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
git-filter-repo==2.47.0
openpyxl==3.1.5
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tzdata==2025.2
workbook==1.1
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
